refactored: set password screen
