+++ lib/ui/screens/security_set_password_screen/widgets/set_password_content/widgets/password_input_field/index.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:mooncake/ui/ui.dart';
+import 'package:mooncake/ui/screens/security_set_password_screen/blocs/export.dart';
 
 /// Represents the input that is used from the user to insert a password.
 class PasswordInputField extends StatelessWidget {
