+++ lib/ui/widgets/password_strength_indicator/index.dart
