--- lib/ui/widgets/security_set_password/export.dart
@@ -1,2 +1,0 @@
-export 'set_password_content.dart';
-export 'saving_password_popup_content.dart';
